feat(checkout): 实现 checkout功能并调整相关服务
- 更新 checkout 服务实现，包括订单创建和支付处理
- 调整 client 服务中的 checkout 接口和路由
- 修改 order服务中的订单列表展示
-优化 auth 服务的 go.mod 和 go.sum 文件
- 移除 checkout 服务中的 dal 初始化

+++ app/order/biz/service/list_order.go
@@ -30,7 +30,7 @@ func (s *ListOrderService) Run(req *order.ListOrderReq) (resp *order.ListOrderRe
 				Cost: oi.Cost,
 				Item: &cart.CartItem{
 					ProductId: oi.ProductId,
-					Quantity:  int32(oi.Quantity),
+					Quantity:  oi.Quantity,
 				},
 			})
 		}
